Version bump to begin Bioc 1.8 DEVEL work.
This version bump is unusual in that it is part of a catch-up version bump.
The project has decided to work with an even/odd scheme.

All versions are x.y.z.  
In release y is even.
In devel y is odd.

This change sets everyone to the next odd y (next as compared to
version in 1.7 release).


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Biostrings@14469 bc3139a8-67e5-0310-9ffc-ced21a209358