Add the alignedPattern() and alignedSubject() getters for extracting the aligned patterns and subjects from a PairwiseAlignments object.
Unlike aligned(pattern()) and aligned(subject()), alignedPattern()
and alignedSubject() do NOT trim globally aligned sequences. The
fact that aligned(pattern()) and aligned(subject()) trim the deleted
extremities of globally aligned sequences as been a nuisance and a
source of complain for many years. The addition of alignedPattern()
and alignedSubject() is a response to this.

The "show" method for PairwiseAlignments objects now uses
alignedPattern() and alignedSubject() internally to display the
untrimmed globally aligned pattern and subject.

The "as.character" method for PairwiseAlignmentsSingleSubject objects
was modified to do as.character(alignedPattern(x)) so does not trim
globally aligned patterns anymore.