Add 'no.init.codon' argument to translate()
By default, translate() assumes that the first codon in a DNA or RNA
sequence is the initiation codon. This means that the "alt_init_codons"
attribute on the supplied 'genetic.code' is used to translate the
alternative initiation codons. This can now be changed by setting the
new 'no.init.codon' argument to TRUE, in which case the "alt_init_codons"
attribute is ignored.