Calls to functions defined in pwalign are now redirected there
The following Biostrings functions are now implemented in the new
pwalign package:
  - pairwiseAlignment
  - PairwiseAlignments
  - PairwiseAlignmentsSingleSubject
  - writePairwiseAlignments
  - alignedPattern
  - alignedSubject
  - insertion
  - deletion
  - unaligned
  - aligned
  - indel
  - nindel
  - nedit
  - pid
  - mismatchTable
  - mismatchSummary
  - compareStrings
  - stringDist
  - nucleotideSubstitutionMatrix
  - errorSubstitutionMatrices
  - qualitySubstitutionMatrices

Note that they are still temporarily defined in Biostrings but now
they just call the corresponding function in pwalign. Since this is a
temporary redirect, the user also gets a warning that tells them to use
the fully qualified name (e.g. pwalign::pairwiseAlignment()) to call
the function.